CODE REFACTORING
