--- app.js
@@ -1,4 +1,0 @@
-const express = require('express');
-const server2 = express();
-
-module.exports = server2;
