+++ index.js
@@ -1,17 +1,18 @@
 const express = require('express');
 const ORDERROUTES = require('./routes/orderRoutes');
+const USERROUTES = require('./routes/usersRoutes');
 const AUTH = require('./routes/authRoutes');
 const morgan = require('morgan');
 const AppError = require('./utils/AppError');
 const errController = require('./controllers/errController');
-const reqController = require('./controllers/errController');
+
 const app = express();
 app.use(express.json());
-// app.use(morgan('common'));
-app.use(morgan('dev'));
+app.use(morgan('common'));
 
 app.use('/api/orders', ORDERROUTES);
 app.use('/api/', AUTH);
+app.use('/api/users', USERROUTES);
 
 app.all('*', (req, res, next) => {
 	next(
