+++ routes/usersRoutes.js
@@ -1,6 +1,9 @@
-const router = require('express ').Router();
+const router = require('express').Router();
 const userController = require('../controllers/usersControllers');
+const auth = require('../controllers/auth');
 
-router.route('/').get(userController);
+router.route('/signup').post(auth.Register);
+
+// router.route('/').get(userController);
 
 module.exports = router;
