--- serverRoom.js
@@ -1,15 +1,0 @@
-const { createServer } = require('http');
-
-class server extends createServer {
-	constructor(app, port) {
-		super();
-		this.port = port;
-		this.app = app;
-
-		createServer(this.app).listen(this.port, () => {
-			console.log(`Node is listening on ${this.port}`);
-		});
-	}
-}
-
-module.exports = server;
